backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.8.0
    # via -r requirements.in
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via
    #   -r requirements.in
    #   httpcore
    #   starlette
    #   watchfiles
astroid==2.11.6
    # via pylint
async-timeout==4.0.2
    # via redis
asyncpg==0.25.0
    # via -r requirements.in
attrs==21.4.0
    # via pytest
babel==2.10.3
    # via -r requirements.in
bandit==1.7.4
    # via -r requirements.in
bcrypt==3.2.2
    # via passlib
billiard==3.6.4.0
    # via celery
black==22.6.0
    # via -r requirements.in
build==0.8.0
    # via pip-tools
celery==5.2.7
    # via -r requirements.in
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage[toml]==6.4.1
    # via pytest-cov
cryptography==37.0.2
    # via python-jose
deprecated==1.2.13
    # via redis
dill==0.3.5.1
    # via pylint
dnspython==2.2.1
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.2.1
    # via pydantic
fastapi==0.78.0
    # via
    #   -r requirements.in
    #   fastapi-jwt-auth
fastapi-jwt-auth==0.5.0
    # via -r requirements.in
flake8==4.0.1
    # via
    #   -r requirements.in
    #   flake8-docstrings
    #   flake8-print
flake8-docstrings==1.6.0
    # via -r requirements.in
flake8-print==5.0.0
    # via -r requirements.in
freezegun==1.2.1
    # via -r requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via bandit
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httptools==0.4.0
    # via uvicorn
httpx==0.23.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via
    #   -r requirements.in
    #   pylint
jinja2==3.1.2
    # via -r requirements.in
kombu==5.2.4
    # via
    #   -r requirements.in
    #   celery
lazy-object-proxy==1.7.1
    # via astroid
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy==0.961
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
orjson==3.7.5
    # via -r requirements.in
packaging==21.3
    # via
    #   build
    #   pytest
    #   redis
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.9.0
    # via black
pbr==5.9.0
    # via stevedore
pep517==0.12.0
    # via build
pip-tools==6.7.0
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.30
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycodestyle==2.8.0
    # via
    #   flake8
    #   flake8-print
pycparser==2.21
    # via cffi
pydantic[email]==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
    #   sqlmodel
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.4.0
    # via flake8
pyhumps==3.7.2
    # via -r requirements.in
pyjwt==1.7.1
    # via fastapi-jwt-auth
pylint==2.14.3
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==3.0.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via freezegun
python-dotenv==0.20.0
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pytz==2022.1
    # via
    #   babel
    #   celery
pyupgrade==2.34.0
    # via -r requirements.in
pyyaml==6.0
    # via
    #   bandit
    #   uvicorn
redis==4.3.4
    # via -r requirements.in
requests==2.28.0
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
rsa==4.8
    # via python-jose
sentry-sdk==1.6.0
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   ecdsa
    #   python-dateutil
    #   python-multipart
smmap==5.0.0
    # via gitdb
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
snowballstemmer==2.2.0
    # via pydocstyle
sqlalchemy==1.4.35
    # via
    #   -r requirements.in
    #   alembic
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a24
    # via sqlmodel
sqlmodel==0.0.6
    # via -r requirements.in
starlette==0.19.1
    # via
    #   -r requirements.in
    #   fastapi
stevedore==3.5.0
    # via bandit
tokenize-rt==4.2.1
    # via pyupgrade
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pep517
    #   pylint
    #   pytest
tomlkit==0.11.0
    # via pylint
types-freezegun==1.1.10
    # via -r requirements.in
types-orjson==3.6.2
    # via -r requirements.in
types-redis==4.3.2
    # via -r requirements.in
typing-extensions==4.2.0
    # via
    #   mypy
    #   pydantic
    #   sqlalchemy2-stubs
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.18.2
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.15.0
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.3
    # via uvicorn
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via
    #   astroid
    #   deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
